Clean up a few things, move reconstructor to its own file
